Explicitly list out the action for each API endpoint.
Instead of relying on path parsing to figure out the action tag for each API
endpoint, we will explicitly enumerate the actions. Less magic and easier for
the end user to figure out which action corresponds to which API endpoint. You
can just `ack` the code to find the action name.