Properly set the read and open timeouts for faraday.
Turns out that faraday is using the Net:HTTP timeout notation, and the
adapters adapt those to whatever backend conneciton is going to be used. In
the case of excon, even though it needs a `read_timeout` option, we still need
to pass `timeout` to the Faraday options. The excon adapter takes care of the
mapping. TMYK