Initial whack at some validations of the URL components.
I'm taking advantage of the fact that we are using Addressable::Templates to
generate our URLs from user supplied paramters. We can identify required
elements `{index}` versus optional elements `{/index}` and validate these
components accordingly.

So if an element is present, it cannot be nil and it cannot be blank. The
`validate_*` methods enforce these rules. An ArgumentError is raised if the
validations fail.