wait for yellow status instead of green
In tests, all we care about is having at least one shard active. There's
no need to wait for a fully replicated index, something that will never
happen with the default settings on a single node cluster.

This is only an issue when testing against non-boxen Elasticsearch. Boxen
sets the default number of replicas to zero, but ES packages downloaded
from the ES website set the default number of replicas to 1.