add a delete_by_query test
The delete_by_query api was untested before, which is probably why nobody
noticed that the query hash argument never worked. Fortunately it
will work on 1.0 because it adds the top level :query element to the
request format.

The test skips testing the query hash argument unless it's running
against 1.0.