Properly encapsulate a Hash into an Array
The ruby `Array()` method can be used to easily wrap the arguments into an
array. When given a Hash, though, it will create an array of [key, value] pairs
not an hash wrappered up in an array [{key => value}]

So we need to change the method handling here to do the right thing.