Support regular scroll queries
We have support for scan queries, but we do not expose the scroll query. The
scroll query allows us to sort results. The Scan query does not allow sorting,
and therefore, it is a bit more efficient than a regular scroll query.

These changes allow us to support both via a Scroller object.