Precompute size and action count for bulk indexing
This change to the bulking indexer will prevent the max request size and the
bulk request size from being exceed. When a bulk indexing action is requested,
the bytesize of that action is computed first. If adding this action to the
queue would cause the `request_size` to be exceeded, then the bulk request is
immediately dispatched and the new action is added as the first action for the
next bulk request.