Adding some exceptions wrappers
This PR is aimed at adding some exception wrappers around common error responses
from Elasticsearch. The problem is that in some of our error handling code we
have lines like this:

```ruby
rescue StandardError => err
  if err.message =~ %r/IndexMissingException|index_not_found_exception/
    # do someting here
  end
```

A better approach is to rescue a named exception and not worry about using a
regular expression on the error message to figure out what the problem actually
is.