Handle difference in ES 5 bulk API behavior
I couldn't find this in the changelogs or API docs, but experimentally ES 5
responds to a bulk "index" operation with no ID with a "index" response instead
of a "create" response.

This also adds awesome_print to help debug output during tests. When you need to
use it `require "awesome_print"` and then use `ap`.