Move build prerequisites to install phase
Doing multiple steps in the script phase was causing builds to not fail-fast if
Elasticsearch is not started.

However, the bigger problem is that...Elasticsearch wasn't started. I moved the
test for that to run after installing gems so it will have more time to start
up. I have not changed the wget arguments but I did use the long options so it's
easier to understand. If necessary we can increase --waitretry or --tries.