Add VersionSupport class for ES version differences
VersionSupport is pretty basic right now but the idea is to have a place to
stash version-aware helpers. Currently it's used to provide a way to paper over
the difference between `text` and `keyword` fields in ES 5+ and `string` fields
in ES 2.

Supposedly ES 5 can take a legacy `string` field and do the right thing with it,
but there are severe limitations, such as not being able to handle an a `string`
field with `term_vectors` enabled. :(

See: https://github.com/elastic/elasticsearch/issues/21356