Update analyze to use 2.0+ API
Elasticsearch 1 and prior's analyze API allowed you to pass plain text to
analyze. This was deprecated in 2.0 and removed in 5.0 in favor of passing a
JSON object with a `text` key.

Since this version of the API works in both ES 2 and 5, I've updated the code to
use the new way in order to fix the failing test.