Fail fast on known invalid indexing directive fields
Pull the different "special" fields that affect indexing into VersionSupport in
order to return a different Hash of parameter name/field name depending on the
client's version.

If a caller attempts to index a document with a known indexing directive field
that is not supported by the version of Elasticsearch, the client will fail fast
and not allow the document to be indexed.

Fixes #178.