Delegate to most appropriate delete_by_query implementation
This repurposes Elastomer::Client#delete_by_query such that it delegates to the
most appropriate delete by query implementation for the version of
Elasticsearch. The old delete_by_query method and class are now known as
app_delete_by_query and AppDeleteByQuery, respectively.

Tests are updated for the API change and documentation has been added for the
new native_delete_by_query implementation and to warn about the behavior change
in the parameters and response format.