Pass in config instead of trying to infer it
ES 2 supports compression but it's not enabled by default. Vice versa,
compression can be disabled for ES 5. So instead of trying to infer the if
compression can be used, tell the client.

Piggybacking this change, I added configuration to allo tell the client what the
ES version is instead of sniffing it, because I was seeing 15-20ms times to do
this check.