Fix issues with generate-rest-api-spec
Previously this didn't truncate the file, so if the newly generated file
was smaller it would create a syntax error.

This also avoids loading anything except for version_support, since that
was the only part needed. To achieve that I switched to using JSON from
the stdlib instead of MutliJson