Unlock faraday_middleware version
This is a followup to https://github.com/github/elastomer-client/pull/309

We can't upgrade faraday to 1.x without also upgrading
faraday_middleware to 1.x, so the previous PR didn't get us all the way
there.

Not much changes in faraday_middleware 1.x except for faraday 1.x
support, so I expect elastomer-client to be fully compatible.

(We'll revisit this once more for faraday 2.0, which drops the
faraday_middleware gem entirely, but that's a problem for another day.)