Fix tests for Faraday 1.0
The adapter is now stored separately (it's been that way for some time,
but we're newly using the new mechanism after upgrading).

Before:

```
<Faraday::RackBuilder:0x00000001237fabb0 @handlers=[ElastomerClient::Middleware::ParseJson, FaradayMiddleware::Gzip, ElastomerClient::Middleware::EncodeJson, ElastomerClient::Middleware::LimitSize, Faraday::Adapter::Test]>
```

After:

```
<Faraday::RackBuilder:0x00000001226b8e38 @adapter=Faraday::Adapter::Test, @handlers=[ElastomerClient::Middleware::ParseJson, FaradayMiddleware::Gzip, ElastomerClient::Middleware::EncodeJson, ElastomerClient::Middleware::LimitSize]>
```

So this commit updates these tests to check that we're using the right
adapter instead of checking that the adapter is in the list of handlers.